Change the way dimensions are used in the table
This modification originates from the issue that manually reordering dimension columns does not change the column order state - it only changes the order of entries in the hypercube's qDimensions list. Because of this, it was easier to not treat dimensions as separate additions. Rather custom dimensions are stored alongside default table fields in a new `tableDimensions` property to keep track of column order and stored manipulations for later.

Distinction between default table fields and custom dimensions is made with the `isField` and `isDimension` property on the hypercube dimension's definition object. To utilize the dimension's elements for parsing expressions the `dimension` property is stored there as well.

Along with the rewrite a bunch of code is removed that would patch object properties when no changes are made. This decreases the amount of patch data communicated to the Engine.

In addition a new 'Remove all fields' action is added to the list of 'Remove column' actions to remove all default table fields.