Add custom footnote showing the total number of rows
This is done using a custom template element because of limitations in recalculating the total number of rows in an expression. Using the hypercube's size on the object's layout is always correct, but this is only available after the table is already rendered. Consequently the layout's data is used in a custom footnote footer outside of the embedded visualization.