Introduce the table profile view
The table profile view presents metadata for the table's fields. You can switch between the table inspector and the table preview through an action in the context menu or using the button in the extension's footer. The table profile view has its own FSM states and transitions.

Converting the extension to a straight table with the table view active is supported.